backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
google-genai==1.0.0
reportlab==4.2.0
python-dotenv==1.0.1
